src/auth/config.py
from dotenv import load_dotenv
import os

# Charger les variables d'environnement
load_dotenv()

# Configuration des variables
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))
