requirements.txt
fastapi
passlib[bcrypt]
python-jose
uvicorn
python-multipart
bcrypt==3.2.0
python-dotenv